Brazil admin 2 update
Pcodes for admin 2 in Brazil (Municipio) have been created. Geometry file to visualize admin 2 levels in Brazil have always been created and added.

The file is larger than most other files (2.83mb) so further compression should be looked at.

Tested and adm2 visualizes correctly.

This is part of issue #66 , but does not completely resolve it as the file size is too large still.